Move the V1 (default) API to it's own folder structure in prep for V2. V2 is not deployed to all states and has new tables. This will allow developers to use either version depending on the conditions.

+++ src/OpenMetrc.Builder/Models/Harvest.cs
@@ -2,7 +2,7 @@
 
 public class Harvest
 {
-    [Required] public int Id { get; set; }
+    [Required] public long Id { get; set; }
 
     public string? Name { get; set; }
     public string? HarvestType { get; set; }
