+++ src/OpenMetrc.Builder/Models/LocationType.cs
@@ -2,7 +2,7 @@
 
 public class LocationType
 {
-    [Required] public int Id { get; set; }
+    [Required] public long Id { get; set; }
 
     public string? Name { get; set; }
     public bool? ForPlantBatches { get; set; }
