+++ src/OpenMetrc.Builder/Models/PatientRegistrationLocation.cs
@@ -2,7 +2,7 @@
 
 public class PatientRegistrationLocation
 {
-    [Required] public int Id { get; set; }
+    [Required] public long Id { get; set; }
 
     public string? Name { get; set; }
 }
