+++ src/OpenMetrc.Builder/Models/PlantBatch.cs
@@ -2,7 +2,7 @@
 
 public class PlantBatch
 {
-    [Required] public int Id { get; set; }
+    [Required] public long Id { get; set; }
 
     public string? Name { get; set; }
     public int? PlantBatchTypeId { get; set; }
