+++ src/OpenMetrc.Builder/Models/RequiredLabTestBatch.cs
@@ -7,7 +7,7 @@ public class RequiredLabTestBatch
         LabTestBatchName = string.Empty;
     }
 
-    public int PackageId { get; set; }
-    public int LabTestBatchId { get; set; }
+    public long PackageId { get; set; }
+    public long LabTestBatchId { get; set; }
     public string LabTestBatchName { get; set; }
 }
